+++ tests/suites/tester.js
@@ -43,6 +43,12 @@ casper.thenOpen('tests/site/index.html', function() {
     t.comment('Tester.assertTrue()');
     t.assertTrue(true, 'Tester.assertTrue() works as expected [alias]');
 
+    t.comment('Tester.assertTruthy()');
+    t.assertTruthy('1', 'Tester.assertTruthy() works as expected');
+
+    t.comment('Tester.assertFalsy()');
+    t.assertFalsy('0', 'Tester.assertFalsy() works as expected');
+
     t.comment('Tester.assertNot()');
     t.assertNot(false, 'Tester.assertNot() works as expected');
 
@@ -205,5 +211,5 @@ casper.then(function() {
 });
 
 casper.run(function() {
-    t.done(56);
+    t.done(58);
 });